requirements.txt
Django==1.11.3
dj-database-url==0.4.2
psycopg2==2.7.3
gunicorn==19.7.1
whitenoise==3.3.0
django-material==1.0.0
django-jquery==3.1.0
django-rank-query==0.1.0